refactor(cli): extract group dispatchers into handlers/
cli.ts now holds runCli + TOP_USAGE + defaultEnv only. Each group
(mcp, backup, inspect) and the special mcp-serve flow get their own
file under handlers/, owning the group usage string and verb switch.
No behavior change.
